requirements.txt
# Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
openai>=1.12.0
httpx>=0.26.0
python-dotenv>=1.0.0
pydantic>=2.6.0
PyJWT>=2.8.0

# Queue & Workers
redis>=5.0.0
rq>=1.16.0

# SSE
sse-starlette>=2.0.0

# Database (using SQLite for MVP, can switch to Postgres later)
aiosqlite>=0.19.0

# Composio SDK (tools + OpenAI provider)
composio
composio_openai
